+++ changelog.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## [5.18.1 (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v5.18.0...main)
+## [5.18.2 (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v5.18.1...main)
 
 > \[!IMPORTANT\]
 > This version is not released yet and is under active development.
+
+## [5.18.1 (2024-08-02)](https://github.com/kdeldycke/meta-package-manager/compare/v5.18.0...v5.18.1)
 
 ## [5.18.0 (2024-08-02)](https://github.com/kdeldycke/meta-package-manager/compare/v5.17.0...v5.18.0)